Resume-screening-Streamlit-main/requirements.txt
altair==4.2.0
anyio==3.6.1
argcomplete==1.10.3
asttokens==2.0.8
attrs==22.1.0
backcall==0.2.0
bcrypt==4.0.1
beautifulsoup4==4.8.2
blinker==1.5
cachetools==5.2.0
certifi==2022.9.24
cffi==1.15.1
chardet==3.0.4
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
commonmark==0.9.1
compressed-rtf==1.0.6
cryptography==38.0.1
debugpy==1.6.3
decorator==5.1.1
dnspython==2.2.1
doc2text==0.2.4
docopt==0.6.2
docx2txt==0.8
ebcdic==1.1.1
ecdsa==0.18.0
email-validator==1.3.0
entrypoints==0.4
executing==1.1.1
extra-streamlit-components==0.1.56
extract-msg==0.28.7
fastapi==0.85.1
filelock==3.8.0
future==0.18.2
gitdb==4.0.9
GitPython==3.1.29
greenlet==1.1.3.post0
h11==0.14.0
htbuilder==0.6.0
httptools==0.5.0
huggingface-hub==0.10.1
idna==3.4
IMAPClient==2.1.0
importlib-metadata==5.0.0
ipykernel==6.16.1
ipython==8.5.0
iteration-utilities==0.11.0
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.16.0
jupyter_client==7.4.3
jupyter_core==4.11.2
lxml==4.9.1
MarkupSafe==2.1.1
matplotlib-inline==0.1.6
mime==0.1.0
nest-asyncio==1.5.6
nltk==3.7
numpy==1.23.4
olefile==0.46
orjson==3.8.0
packaging==21.3
pandas==1.5.1
parso==0.8.3
passlib==1.7.4
pickleshare==0.7.5
Pillow==9.2.0
pipreqs==0.4.11
prompt-toolkit==3.0.31
protobuf==3.20.3
psutil==5.9.3
psycopg2-binary==2.9.4
pure-eval==0.2.2
pyarrow==9.0.0
pyasn1==0.4.8
pycparser==2.21
pycryptodome==3.15.0
pydantic==1.10.2
pydeck==0.8.0b4
Pygments==2.13.0
PyJWT==2.6.0
Pympler==1.0.1
pyparsing==3.0.9
PyPDF2==2.11.1
pyrsistent==0.18.1
pytesseract==0.3.10
python-dateutil==2.8.2
python-docx==0.8.11
python-dotenv==0.21.0
python-jose==3.3.0
python-multipart==0.0.5
python-pptx==0.6.21
pytz==2023.3
PyYAML==6.0
pyzmq==24.0.1
regex==2022.9.13
requests==2.28.1
rich==12.6.0
rsa==4.9
scikit-learn==1.1.2
scipy==1.9.3
semver==2.13.0
six==1.12.0
sklearn==0.0
smmap==5.0.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
SpeechRecognition==3.8.1
SQLAlchemy==1.4.42
st-annotated-text==3.0.0
stack-data==0.5.1
starlette==0.20.4
streamlit==1.13.0
streamlit-authenticator==0.2.1
streamlit-book==0.7.5
streamlit-camera-input-live==0.2.0
streamlit-embedcode==0.1.2
streamlit-extras==0.1.5
streamlit-keyup==0.1.9
streamlit-option-menu==0.3.2
textract==1.6.5
threadpoolctl==3.1.0
tokenizers==0.13.1
toml==0.10.2
toolz==0.12.0
torch==1.12.1
tornado==6.2
tqdm==4.64.1
traitlets==5.5.0
transformers==4.23.1
typing_extensions==4.4.0
tzdata==2022.5
tzlocal==4.2
ujson==5.5.0
urllib3==1.26.12
uvicorn==0.18.3
validators==0.20.0
Wand==0.6.10
watchdog==2.1.9
watchfiles==0.17.0
wcwidth==0.2.5
websockets==10.3
xlrd==1.2.0
XlsxWriter==3.0.3
yarg==0.1.9
zipp==3.9.0
